Player movement overhaul. Merged Tilt++ and WeaponSway natively into TCOTD actor classes with highly customized tweaks specifically designed for TCOTD's immersion and atmosphere.
As an added bonus, there is also Blood-style jumping, landing and crouching sway which is not even available publicly in WeaponSway - it will make its debut for the first time in TCOTD:A as I don't plan to release this version of WeaponSway for universal usage.

(Yes, various settings previously exposed to the player in the menu from those 2 mods are intentionally hardcoded now, I don't want players ruining the experience. The only customizability I made available is a simple On/Off toggle)

Movement speeds are not 1:1 with previous TCOTD movement but based on my initial testing, I feel that it comes close enough. Can be further adjusted as needed.