Add files via upload
Changed the slime sector so it's fully connected w/o portals. The floors are all 3D floors. Also ensured the switch in that sector is only activated at the exact height. Moreover, added a couple more little details and changed the exit bit slightly so it's connected by a line portal to prevent it from going over the slime sector. Also converted it to UDMF my end and may continue more work on it.