STR01 has it's lighting finished and changed the floor...
Texture in the mainbase to make it consistent with MAP99. Speaking of which, converted MAP99 into UDMF, and sorted most of the lighting out in tiers 1 & 2, and make a wish in the water fountains in the same tiers, or alternatively, admire the reflection I added into the water. :P

Will continue this at some point.

Moreover, the scripting is giving me grief. Will need assistance.