[TCOTD2] Fixed: spawn-dead monsters counted twice towards the total-monsters counter Once by the call to A_ChangeCountFlags() (which is now removed. I thought it was needed in such a scenario), and another upon being resurrected. This resulted in not being able to achieve a 100% kill count at the end of the game, even if all monsters in the level are dead.
The switch from Die() to A_Die() is because Die() doesn't reduce the monster's health to 0 (I don't
know why I thought it would do, since it's not its job to do so), which caused A_ChangeCountFlags()
to up the total-monsters counter by one.

A_ChangeCountFlags() is still unneeded, hence its removal, as all it'd be doing is set the
COUNTKILL flag, which the resurrection code already takes care of.