[TCotD2] Added displacement maps
This is partially experimental, as I do not want to add these maps to all textures, but in some cases, there is a huge benefit, for example in the mortuary. I am going to add a few more where it makes sense, feel free to support me where you think it improves the visuals. By the way, the included script writes the displacement gl file. No need to do that, I will take care of it.