[TCOTD2] Pistol code rework This is to rid the weapon of any possible issues related to reloading, like the one it had where it automatically reloaded after firing the last bullet the player had.
Additionally, the reload sequence has been changed so that it's always the same in terms of length and animation, regardless of how many bullets are needed to fully reload the weapon. The argument here is that player character is using a speedloader for reloading, instead of reloading one bullet at a time.