[TCOTD2] Fix giant TCOTD2 logo in the menu.
Files not meant to be loaded into the game should not be placed in the active namespaces. What was happening was that GZDoom was trying to load M_DOOM.psd (as much as that doesn't make sense).

To avoid future problems, I'm moving every "graphic source" file (including the font TTFs/OTFs) into a PSD folder